Updated Home page to include four color-coded value boxes representing current QC status for the individual instruments.
The third colum over is the individual instrument tabs for the respective detectors, color coded red fail today, orange fail in the week, green pass.

The fourth column is the three month history for all instruments.

Sidebar and links to core to the website and last upload date for context added.