Thanksgiving Break Update.
The 4L data transfer failed due to a shut-down on the 26th. I think the other transfers failed due to the FlagRemoval.log still being present in the system? (unclear). But the 27th should have had enough data that morning to forward but didn't.

Also, Rstudio project was not open to correct location on opening, unclear if connected